+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 
 ### New
 
+
+- Add New Relic Scaler ([#2387](https://github.com/kedacore/keda/pull/2387))
+
 - Add ActiveMQ Scaler ([#2305](https://github.com/kedacore/keda/pull/2305))
 
 - TODO ([#XXX](https://github.com/kedacore/keda/pull/XXX))